Adds CLI for update, renames action-factory > prompt
